First WAT module
Create a WAT module with the module keyword. Import "env" prepares
the module to accept an import object with a function property
called print_string with a parameter representing the length of
the string.

We also set up another import object property representing the
memory buffer, called buffer, which is one page (i.e. chunk) of
memory in size, which is 64kb in Wasm.

Next, we import a global variable from javascript called
start_string, representing the starting memory location. In our
Js we set that to 100, so as not to reach our 64kb 1-page limit.

string_len is a const defined in Wasm for the length of the
string, which is 12.

We then define our string in linear memory by passing the
location using global.get $start_string, then passing the
string itself.

We can then define and export our function, which calls
the imported function with the length of the string as
an argument (and I assume the location is already on the
stack for it to use).