Create Javascript file to execute generated Wasm module
wat2wasm helloworld.wat generates a wasm file, which we now can
call with a Js file.

We read the wasm file with readFileSync, and declare variables
pertaining to the helloWorld function to be called, the
startStringIndex pertaining to the memory location to be used,
and memory pertaining to the 64kb single-page linear Wasm memory
to be used.

Next we store these in an import object, with a top-level property
of env as expected by the Wasm file, within which we have the
expected buffer of memory, the start_string memory location and
the print_string function that creates an array of memory using
the buffer, the byteOffset and the length (see MDN).

This is then passed to a TextDecoder which decodes the buffer
into a string to be console logged.

Finally, the wasm code is instiantiated along with the import
object and all its properties matching the values to be imported
into the Wasm instance.

With this, we grab the helloworld property from the instance
exports and assign it to the Javascript variable, which we
can then call, successfully console logging hello world.