Write factorial in wat to demonstrate loop and block
First we import the log function from Js which will allow
us to print the factorial result.

Next we define a function loop_test to be exported, with
param and result (return) n of type i32.

Local variables $i and $factorial are declared, the latter
set to an initial value of 1.

A loop expression labelled $continue is opened, with a
block statement labelled $break opened inside it.

$i is then set to $i + 1. $factorial is set to $i * $factorial.
Our print function, $log, is then called with params $i and
$factorial.

br_if determines whether to break out of the loop or return to
the top. If $i has reached the value of $n, we break. If not,
we loop more.