Write is_prime function
First we check if the number is equal to 1 or 2 and return
accordingly. Then we enter a block from which we break if at any
point the number is determined not to be prime. First we check
for evenness because any even number other than the previously
checked 2 is not prime. If even, break. We begin a loop at 3,
incrementing by 2. If we have reached n without breaking, we can
return 1 (true) because it must be prime since none of our
non-prime checks passed along the way. We check if n is a multiple
of the incrementing counter, and break if so. If not, continue
looping until either the non-prime checks pass or the counter
reaches n. If broken, 0 is returned to represent false for primality.