Compare a 4 million loop function in native wasm versus Js export
First we import the external Js call to wasm. Then we define our
internal call that simply adds 1 to the parameter. Next we declare
a function to export which loops an increment until 4 million.
Finally, we declare a function to export which makes a callto the
external Js function, doing the same 4 million increment.

In Js, we initialize our increment counter, our import object which
includes the external_call increment function, we instantiate Wasm,
pull wasm_call and js_call out of the exports, and make a call to
each, wrapped by process.hrtime.bigint() for nanosecond accurate
performance comparisons. Wasm call rounds out to ~10ms vs ~34ms
for Js.